finish lab1:TinyKV part

+++ tinykv/kv/transaction/commands/rollback.go
@@ -47,7 +47,6 @@ func rollbackKey(key []byte, txn *mvcc.MvccTxn, response interface{}) (interface
 		zap.Uint64("startTS", txn.StartTS),
 		zap.String("key", hex.EncodeToString(key)))
 
-	panic("rollbackKey is not implemented yet")
 	if lock == nil || lock.Ts != txn.StartTS {
 		// There is no lock, check the write status.
 		existingWrite, ts, err := txn.CurrentWrite(key)
@@ -61,33 +60,68 @@ func rollbackKey(key []byte, txn *mvcc.MvccTxn, response interface{}) (interface
 		// If the key has already been committed. This should not happen since the client should never send both
 		// commit and rollback requests.
 		// There is no write either, presumably the prewrite was lost. We insert a rollback write anyway.
-		if existingWrite == nil {
-			// YOUR CODE HERE (lab1).
 
-			return nil, nil
-		} else {
-			if existingWrite.Kind == mvcc.WriteKindRollback {
-				// The key has already been rolled back, so nothing to do.
-				return nil, nil
-			}
+		//if existingWrite == nil {
+		//	// YOUR CODE HERE (lab1).
+		//	// prewrite丢失，也需要插入一个回滚记录
+		//	write := mvcc.Write{StartTS: txn.StartTS, Kind: mvcc.WriteKindRollback}
+		//	txn.PutWrite(key, txn.StartTS, &write)
+		//	txn.DeleteLock(key)
+		//	return nil, nil
+		//} else {
+		//	if existingWrite.Kind == mvcc.WriteKindRollback {
+		//		// The key has already been rolled back, so nothing to do.
+		//		return nil, nil
+		//	}
+		//
+		//	// The key has already been committed. This should not happen since the client should never send both
+		//	// commit and rollback requests.
+		//	// 错误情况：key已经被提交
+		//	err := new(kvrpcpb.KeyError)
+		//	err.Abort = fmt.Sprintf("key has already been committed: %v at %d", key, ts)
+		//	respValue := reflect.ValueOf(response)
+		//	reflect.Indirect(respValue).FieldByName("Error").Set(reflect.ValueOf(err))
+		//	return response, nil
+		//}
 
-			// The key has already been committed. This should not happen since the client should never send both
-			// commit and rollback requests.
+		// 1.存在当前事务的提交记录
+		// 错误情况：不应该发生
+		if existingWrite != nil && existingWrite.Kind != mvcc.WriteKindRollback && existingWrite.StartTS == txn.StartTS {
 			err := new(kvrpcpb.KeyError)
 			err.Abort = fmt.Sprintf("key has already been committed: %v at %d", key, ts)
 			respValue := reflect.ValueOf(response)
 			reflect.Indirect(respValue).FieldByName("Error").Set(reflect.ValueOf(err))
 			return response, nil
 		}
+		// 2.存在当前事务的回滚记录
+		if existingWrite != nil && existingWrite.Kind == mvcc.WriteKindRollback && existingWrite.StartTS == txn.StartTS {
+			return nil, nil
+		}
+		// 3.不存在当前事务的提交/回滚记录
+		// prewrite丢失，也需要插入一个回滚记录
+		write := mvcc.Write{StartTS: txn.StartTS, Kind: mvcc.WriteKindRollback}
+		txn.PutWrite(key, txn.StartTS, &write)
+		return response, nil
+	}
+	// 检查 key 是否被其他事务锁住
+	if lock.Ts > txn.StartTS {
+		err := new(kvrpcpb.KeyError)
+		err.Abort = fmt.Sprintf("key {%v} has already been locked by other transaction with ts %d", key, txn.StartTS)
+		respValue := reflect.ValueOf(response)
+		reflect.Indirect(respValue).FieldByName("Error").Set(reflect.ValueOf(err))
+		return response, nil
 	}
 
-	if lock.Kind == mvcc.WriteKindPut {
-		txn.DeleteValue(key)
+	if lock.Ts == txn.StartTS {
+		// 删除值
+		if lock.Kind == mvcc.WriteKindPut {
+			txn.DeleteValue(key)
+		}
+		// 写入回滚记录，并删除锁
+		write := mvcc.Write{StartTS: txn.StartTS, Kind: mvcc.WriteKindRollback}
+		txn.PutWrite(key, txn.StartTS, &write)
+		txn.DeleteLock(key)
 	}
-
-	write := mvcc.Write{StartTS: txn.StartTS, Kind: mvcc.WriteKindRollback}
-	txn.PutWrite(key, txn.StartTS, &write)
-	txn.DeleteLock(key)
 
 	return nil, nil
 }
